Do at least some minimal separation
